+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue28968.cs
@@ -2,7 +2,7 @@ using NUnit.Framework;
 using UITest.Appium;
 using UITest.Core;
 
-namespace Microsoft.Maui.TestCases.Tests.Tests.Issues;
+namespace Microsoft.Maui.TestCases.Tests.Issues;
 
 public class Issue28968 : _IssuesUITest
 {
