+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue29588.cs
@@ -2,7 +2,7 @@ using NUnit.Framework;
 using UITest.Appium;
 using UITest.Core;
 
-namespace Microsoft.Maui.TestCases.Tests.Tests.Issues;
+namespace Microsoft.Maui.TestCases.Tests.Issues;
 
 internal class Issue29588 : _IssuesUITest
 {
